+++ tests/UnitTests/MobileDetect/PreparingTheRegexTest.php
@@ -25,4 +25,26 @@ class PreparingTheRegex extends \PHPUnit_Framework_TestCase
             $m->invoke($detect, '^yo[VER]and[MODEL]!$')
         );
     }
+
+    
+
+    protected function assertPattern($pattern, $name)
+    {
+        $detect = new MobileDetect();
+
+        // prepare method so that we can access it
+        $class = new \ReflectionClass(get_class($detect));
+        $method = $class->getMethod('prepareRegex');
+        $method->setAccessible(true);
+        $pattern = $method->invokeArgs($detect, array($pattern));
+
+        $res = @preg_match($pattern, 'garbagestringdoesnotmatter');
+        if ($res === false) {
+            $this->fail(
+                sprintf('Not a valid regex: %s in %s [REGEX FAIL]', $pattern, $name)
+            );
+        }
+        //this is for assertion counts
+        $this->assertInternalType('integer', $res);
+    }
 }
